Allow adding a standby node from an exiting PGDATA directory.
This allows users to use their favorite backup and restore options when
preparing to add a standby node to pg_auto_failover.

We still have a restriction that we don't know how to accept an already
running standby: we would need to check the primary conninfo authentication
options in use, the replication slot, etc. So we still bail out in that
case.